refactor(dev): claim + claim-staleness become re-export shims over the castle engine
apps/dev/src/core/claim.ts and claim-staleness.ts are now 2-line-style
re-export shims from @reddb-io/red-castle/engine (the runner-detection
pattern), so all 14+ dev call sites — process-issue lifecycle, boot
sweeps, requeue, wire/boot, the MCP claim tools, claim-hygiene probes —
run on the single castle implementation without an import-line change.
claim-staleness crosses whole (ADR 0113 module map: castle substrate);
its env/config resolution was already injection-based.

tests/claim-wire-compat.test.ts imports the pinned wire fixtures from the
engine and the parser through the shim, asserting both fixture parity and
referential identity (the shim re-exports the engine's functions, not a
fork). The existing dev claim suites keep running through the shim as the
regression net.
